Added fixture for sql_preprocessor and fixed some issues
